Now keeps almost all auto-generated code from compiling __builtin__.act in __builtin__.c and h. Code previously moved to env.c and h is retained in __builtin__ files
