+++ builtin/builtin.h
@@ -12,6 +12,7 @@
 #include <string.h>
 #include <math.h>
 #include <ctype.h>
+#include <stdatomic.h>
 
 #include "common.h"
 
@@ -43,15 +44,12 @@ struct $action;
 struct $mut;
 struct $pure;
 struct $Cont;
-struct B_Msg;
 
 typedef struct $proc *$proc;
 typedef struct $action *$action;
 typedef struct $mut *$mut;
 typedef struct $pure *$pure;
 typedef struct $Cont *$Cont;
-typedef struct B_Msg *B_Msg;
-
 
 enum $RTAG { $RDONE, $RFAIL, $RCONT, $RWAIT };
 typedef enum $RTAG $RTAG;
@@ -72,6 +70,17 @@ typedef struct $R $R;
 #define $R_WAIT(cont, value)    ($R){$RWAIT, (cont), (value)}
 
 /////////////////////////////////////////////////////////
+
+// And the following from rts.h /////////////////////////
+
+struct $Actor;
+struct $Catcher;
+typedef struct $Actor *$Actor;
+typedef struct $Catcher *$Catcher;
+
+#define $Lock                   volatile atomic_flag
+
+///////////////////////////////////////////////////////////
 
 #include "__builtin__.h"
 #include "class_hierarchy.h"
@@ -100,6 +109,7 @@ typedef struct $R $R;
 #include "exceptions.h"
 #include "function.h"
 #include "builtin_functions.h"
+#include "env0.h"
 #include "staticWitnesses.h"
