+++ builtin/env0.h
@@ -1,0 +1,8 @@
+#pragma once
+
+B_Env B_EnvG_newactor (B_WorldAuth, B_list);
+B_WorldAuth B_WorldAuthG_new();
+B_NoneType B_WorldAuthD___init__ (B_WorldAuth self);
+
+
+
